Refactoring based on code review comments
