+++ package.json
@@ -41,6 +41,7 @@
   ],
   "dependencies": {
     "@ant-design/colors": "^6.0.0",
+    "@aws-amplify/core": "^3.8.21",
     "@babel/plugin-transform-runtime": "^7.9.0",
     "@babel/polyfill": "^7.8.7",
     "@next/bundle-analyzer": "^10.0.9",
